+++ tests/storage/utils.py
@@ -1,19 +1,15 @@
 import ast
 import logging
 import shlex
-from collections.abc import Generator
 from contextlib import contextmanager
 
 import pytest
 import requests
-from kubernetes.dynamic import DynamicClient
-from ocp_resources.cluster_role import ClusterRole
 from ocp_resources.config_map import ConfigMap
 from ocp_resources.daemonset import DaemonSet
 from ocp_resources.datavolume import DataVolume
 from ocp_resources.hostpath_provisioner import HostPathProvisioner
 from ocp_resources.resource import Resource
-from ocp_resources.role_binding import RoleBinding
 from ocp_resources.route import Route
 from ocp_resources.service import Service
 from ocp_resources.storage_class import StorageClass
@@ -160,91 +156,6 @@ class HttpService(Service):
 
 def get_file_url_https_server(images_https_server, file_name):
     return f"{images_https_server}{Images.Cirros.DIR}/{file_name}"
-
-
-@contextmanager
-def create_cluster_role(
-    client: DynamicClient, name: str, api_groups: list[str], verbs: list[str], permissions_to_resources: list[str]
-) -> Generator:
-    """
-    Create cluster role
-    """
-    with ClusterRole(
-        client=client,
-        name=name,
-        rules=[
-            {
-                "apiGroups": api_groups,
-                "resources": permissions_to_resources,
-                "verbs": verbs,
-            },
-        ],
-    ) as cluster_role:
-        yield cluster_role
-
-
-@contextmanager
-def create_role_binding(
-    client: DynamicClient,
-    name: str,
-    namespace: str,
-    subjects_kind: str,
-    subjects_name: str,
-    role_ref_kind: str,
-    role_ref_name: str,
-    subjects_namespace: str | None = None,
-    subjects_api_group: str | None = None,
-) -> Generator:
-    """
-    Create role binding
-    """
-    with RoleBinding(
-        client=client,
-        name=name,
-        namespace=namespace,
-        subjects_kind=subjects_kind,
-        subjects_name=subjects_name,
-        subjects_api_group=subjects_api_group,
-        subjects_namespace=subjects_namespace,
-        role_ref_kind=role_ref_kind,
-        role_ref_name=role_ref_name,
-    ) as role_binding:
-        yield role_binding
-
-
-@contextmanager
-def set_permissions(
-    client: DynamicClient,
-    role_name: str,
-    role_api_groups: list[str],
-    verbs: list[str],
-    permissions_to_resources: list[str],
-    binding_name: str,
-    namespace: str,
-    subjects_name: str,
-    subjects_kind: str = "User",
-    subjects_api_group: str | None = None,
-    subjects_namespace: str | None = None,
-) -> Generator:
-    with create_cluster_role(
-        client=client,
-        name=role_name,
-        api_groups=role_api_groups,
-        permissions_to_resources=permissions_to_resources,
-        verbs=verbs,
-    ) as cluster_role:
-        with create_role_binding(
-            client=client,
-            name=binding_name,
-            namespace=namespace,
-            subjects_kind=subjects_kind,
-            subjects_name=subjects_name,
-            subjects_api_group=subjects_api_group,
-            subjects_namespace=subjects_namespace,
-            role_ref_kind=cluster_role.kind,
-            role_ref_name=cluster_role.name,
-        ):
-            yield
 
 
 def get_importer_pod(
